move isunsignedinteger trait
