+++ math/src/field/traits.rs
@@ -1,5 +1,5 @@
 use super::{element::FieldElement, errors::FieldError};
-use crate::{errors::CreationError, unsigned_integer::u32_word::traits::IsUnsignedInteger};
+use crate::{errors::CreationError, unsigned_integer::traits::IsUnsignedInteger};
 use core::fmt::Debug;
 
 /// Represents different configurations that powers of roots of unity can be in. Some of these may
