+++ math/src/unsigned_integer/u32_word/mod.rs
@@ -1,3 +1,2 @@
 pub mod element;
 pub mod montgomery;
-pub mod traits;
